Move std_uri_template.dart into lib folder This is necessary so that the package can be used after publishing.
